melhorarando o Shift-And com '__int128_t'

+++ include/calcula_casamento.h
@@ -7,6 +7,5 @@
 
 int forcaBruta (list*, list*); // solucao usando forca bruta
 int solucaoBMH (list*, list*, int*); // solucao usando o metodo de Boyer-Moore-Horspool
-int solucaoShiftAnd (list*, list*, unsigned long*); // solucao usando o metodo shift-and
-
+int solucaoShiftAnd (list*, list*, __int128_t*); // solucao usando o metodo shift-and
 #endif
